Changes based on feddback.

+++ src/vm/eventpipe.h
@@ -255,7 +255,7 @@ public:
     static EventPipeSessionID Enable(
         LPCWSTR strOutputPath,
         uint32_t circularBufferSizeInMB,
-        uint32_t profilerSamplingRateInNanoseconds,
+        uint64_t profilerSamplingRateInNanoseconds,
         const EventPipeProviderConfiguration *pProviders,
         uint32_t numProviders,
         EventPipeSessionType sessionType,
